requirements.txt
# Web framework
fastapi[standard]==0.128.0
uvicorn==0.40.0

# Database
sqlmodel==0.0.31
asyncpg==0.31.0
alembic==1.18.1

# Cache
redis==7.1.0
hiredis==3.3.0

# Task queue
celery==5.6.2
flower==2.0.1

# Authentication
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.0.1

# Email
fastapi-mail==1.6.1

# SMS
twilio==9.10.0

# Validation & settings
pydantic-settings==2.12.0
pydantic-extra-types==2.11.0
email-validator==2.3.0

# API docs
scalar_fastapi==1.6.0

# Async
asgiref==3.11.1

# Utilities
python-dotenv==1.2.1
python-multipart==0.0.21
Jinja2==3.1.6
itsdangerous==2.2.0

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
